+++ app/extensions/auth/__init__.py
@@ -4,4 +4,4 @@ Auth extension
 ==============
 """
 
-from .oauth22 import OAuth2Provider
+from .oauth2 import OAuth2Provider
