+++ app/modules/auth/__init__.py
@@ -3,9 +3,16 @@
 Auth module
 ===========
 """
+from flask_login import current_user
 from app.extensions.api import api_v1
 
+def load_user_from_request(request):
+    """
+    Load user from OAuth2 Authentication header.
+    """
+    user = current_user
 
+    return user
 
 def init_app(app, **kwargs):
     # pylint: disable=unused-argument
@@ -19,7 +26,7 @@ def init_app(app, **kwargs):
     api_v1.add_oauth_scope('auth:write', "Provide write access to auth details")
 
     # Touch underlying modules
-    from . import models2, views, resources  # pylint: disable=unused-variable
+    from . import models, views, resources  # pylint: disable=unused-variable
 
     # Mount authentication routes
     app.register_blueprint(views.auth_blueprint)
