+++ tests/modules/auth/resources/test_token.py
@@ -31,7 +31,7 @@ def test_regular_user_can_retrieve_token(
     }
 
     # Clean up
-    from app.modules.auth.models2 import OAuth2Token
+    from app.modules.auth.models import OAuth2Token
     with db.session.begin():
         OAuth2Token.query.filter(OAuth2Token.access_token == response.json['access_token']).delete()
 
@@ -110,7 +110,7 @@ def test_regular_user_can_refresh_token(
     }
 
     # Clean up
-    from app.modules.auth.models2 import OAuth2Token
+    from app.modules.auth.models import OAuth2Token
     with db.session.begin():
         OAuth2Token.query.filter(
             OAuth2Token.access_token == refresh_token_response.json['access_token']
